Efficient Propagations

They create efficient functions which avoid recalculating some parts.

WaveOpticsPropagation.AngularSpectrum — Function

AngularSpectrum(field, z, λ, L; kwargs...)

Returns a function for efficient reuse of pre-calculated kernels.

See angular_spectrum for the full documentation.

Example

julia> field = zeros(ComplexF32, (4,4)); field[3,3] = 1
1

julia> as, t = AngularSpectrum(field, 100e-9, 632e-9, 10e-6);

julia> as(field)
(ComplexF32[6.519258f-8 - 3.5390258f-7im -2.5097302f-7 + 1.1966712f-6im 0.00041754358f0 - 0.00027736276f0im -2.5097302f-7 + 1.1966712f-6im; -2.5136978f-7 + 1.1966767f-6im 8.540863f-7 - 4.0512546f-6im -0.001423778f0 + 0.00093840604f0im 8.540863f-7 - 4.0512546f-6im; 0.00041754544f0 - 0.0002773702f0im -0.0014237723f0 + 0.0009384034f0im 0.5497784f0 + 0.8353027f0im -0.0014237723f0 + 0.0009384034f0im; -2.5136978f-7 + 1.1966767f-6im 8.540863f-7 - 4.0512546f-6im -0.001423778f0 + 0.00093840604f0im 8.540863f-7 - 4.0512546f-6im], (L = 1.0e-5,))

WaveOpticsPropagation.Fraunhofer — Function

Fraunhofer(U, z, λ, L; skip_final_phase=true)

This returns a function for efficient reuse of pre-calculated kernels. See fraunhofer for the full documentation.

Example

julia> field = zeros(ComplexF32, (256,256)); field[130,130] = 1;

julia> f, t = Fraunhofer(field, 4f-3, 632f-9, 100f-6);

julia> f(field)
(ComplexF32[0.00390625f0 + 0.0f0im 0.0039050735f0 - 9.5864176f-5im … 0.0039015447f0 + 0.00019167061f0im 0.0039050735f0 + 9.586406f-5im; 0.0039050735f0 - 9.5864176f-5im 0.0039015447f0 - 0.00019167061f0im … 0.0039050735f0 + 9.586406f-5im 0.00390625f0 - 1.1641532f-10im; … ; 0.0039015447f0 + 0.0001916705f0im 0.0039050735f0 + 9.586407f-5im … 0.0038874403f0 + 0.00038287952f0im 0.003895666f0 + 0.00028736156f0im; 0.0039050735f0 + 9.5864176f-5im 0.00390625f0 - 1.5902765f-12im … 0.003895666f0 + 0.00028736144f0im 0.0039015447f0 + 0.00019167061f0im], (L = 0.006471681f0,))

julia> t.L / 100f-6
64.71681f0

julia> 4f-3 *  632f-9 * 256 / (100f-6)^2
64.71681f0

WaveOpticsPropagation.ScalableAngularSpectrum — Type

ScalableAngularSpectrum(field, z, λ, L; kwargs...)

Returns the electrical field with physical length L and wavelength λ propagated with the scalable angular spectrum method of plane waves (SAS) by the propagation distance z.

Arguments

field: Input field

z: propagation distance

λ: wavelength of field

L: field size (can be a scalar or a tuple) indicating field size

Keyword Arguments

skip_final_phase=true: avoid multiplying with final phase. This phase is also undersampled.

bandlimit_border=(0.8, 1): apply soft bandlimit instead of hard cutoff

Example

julia> field = zeros(ComplexF32, (256,256)); field[130,130] = 1;

julia> sas, t = ScalableAngularSpectrum(field, 10f-3, 633f-9, 500f-6);

julia> res, t2 = sas(field);

julia> t2
(L = 0.00162048f0,)

julia> t
(L = 0.00162048f0,)

julia> t2.L / 500f-6 # calculate magnification
3.24096f0

julia> 10f-3 * 633f-9 * 256 / 500f-6^2 / 2 # equation from paper
3.2409596f0

References

Rainer Heintzmann, Lars Loetgering, and Felix Wechsler, "Scalable angular spectrum propagation," Optica 10, 1407-1416 (2023)

WaveOpticsPropagation.ShiftedAngularSpectrum — Type

ShiftedAngularSpectrum(field, z, λ, L, α; kwargs...)

Returns a method to propagate the electrical field with physical length L and wavelength λ with the shifted angular spectrum method of plane waves (AS) by the propagation distance z. α should be a tuple containing the offset angles with respect to the optical axis.

Arguments

field: Input field

z: propagation distance. Can be a single number or a vector of zs (Or CuVector). In this case the returning array has one dimension more.

λ: wavelength of field

L: field size (can be a scalar or a tuple) indicating field size

α is the tuple of shift angles with respect to the optical axis.

Keyword Arguments

pad_factor=2: padding of 2. Larger numbers are not recommended since they don't provide better results.

bandlimit=true: applies the bandlimit to avoid circular wraparound due to undersampling of the complex propagation kernel [1]

extract_ramp=true: divides the field by phase ramp exp.(1im * 2π / λ * (sin(α[2]) .* x .+ sin(α[1]) .* y)) and multiplies after propagation the ramp (with new real space coordinates) back to the field

Examples

julia> field = zeros(ComplexF32, (4,4)); field[3,3] = 1

julia> AS, _ = ShiftedAngularSpectrum(field, 100e-6, 633e-9, 100e-6, (deg2rad(10), 0));

julia> AS(field)
(ComplexF32[1.5269792f-5 + 1.7594219f-5im -3.996831f-5 - 7.624799f-5im -0.0047351345f0 + 0.002100923f0im -3.996831f-5 - 7.624799f-5im; -8.294997f-5 - 1.8230454f-5im 0.00028230582f0 + 8.1745195f-5im 0.0051693693f0 - 0.016958509f0im 0.00028230582f0 + 8.1745195f-5im; 0.0029884572f0 + 0.0040671355f0im -0.009686601f0 - 0.014245203f0im -0.82990384f0 + 0.5566719f0im -0.009686601f0 - 0.014245203f0im; -1.4191573f-7 + 9.41665f-5im 2.111472f-5 - 0.00031620878f0im -0.017670793f0 - 0.0014212304f0im 2.111472f-5 - 0.00031620878f0im], (L = 0.0001, shift = (0.17632698070846498, 0.0)))

References

Matsushima, Kyoji. "Shifted angular spectrum method for off-axis numerical propagation." Optics Express 18.17 (2010): 18453-18463.

(shifted_as::ShiftedAngularSpectrum)(field)

Uses the struct to efficiently store some pre-calculated objects. Propagate the field.

Propagation

Those were merely for testing purposes.

WaveOpticsPropagation.fraunhofer — Function

fraunhofer(field, z, λ, L)

Returns the electrical field with physical length L and wavelength λ propagated with the Fraunhofer propagation (a single FFT) by the propagation distance z. This is based on a far field approximation z >> λ

This method is efficient but to save memory and avoiding recalculating some arrays (such as the phase kernel), see Fraunhofer.

Arguments

field: Input field

z: propagation distance

λ: wavelength of field

L: field size indicating field size

Keyword Arguments

skip_final_phase=true skip the final phase which is multiplied to the propagated field at the end

Example

julia> field = zeros(ComplexF32, (256,256)); field[130,130] = 1;

julia> res, t = fraunhofer(field, 4f-3, 632f-9, 100f-6)
(ComplexF64[0.00390625 + 0.0im 0.003905073506757617 - 9.586417581886053e-5im … 0.003901544725522399 + 0.0001916706096380949im 0.003905073506757617 + 9.58640594035387e-5im; 0.003905073506757617 - 9.586417581886053e-5im 0.003901544725522399 - 0.0001916706096380949im … 0.003905073506757617 + 9.58640594035387e-5im 0.00390625 - 1.1641532182693481e-10im; … ; 0.003901544725522399 + 0.00019167049322277308im 0.003905073506757617 + 9.586406667949632e-5im … 0.003887440310791135 + 0.0003828795161098242im 0.0038956659846007824 + 0.00028736155945807695im; 0.003905073506757617 + 9.586417581886053e-5im 0.00390625 - 1.5902765215791703e-12im … 0.0038956659846007824 + 0.0002873614430427551im 0.003901544725522399 + 0.0001916706096380949im], (L = 0.006471681f0,))

julia> t.L / 100f-6
64.71681f0

julia> 4f-3 *  632f-9 * 256 / (100f-6)^2
64.71681f0

WaveOpticsPropagation.angular_spectrum — Function

angular_spectrum(field, z, λ, L; kwargs...)

Returns the electrical field with physical length L and wavelength λ propagated with the angular spectrum method of plane waves (AS) by the propagation distance z.

This method is efficient but to avoid recalculating some arrays (such as the phase kernel), see AngularSpectrum.

Arguments

field: Input field

z: propagation distance. Can be a single number or a vector of zs (Or CuVector). In this case the returning array has one dimension more.

λ: wavelength of field

L: field size (can be a scalar or a tuple) indicating field size

Keyword Arguments

padding=true: applies padding to avoid convolution wraparound

pad_factor=2: padding of 2. Larger numbers are not recommended since they don't provide better results.

bandlimit=true: applies the bandlimit to avoid circular wraparound due to undersampling of the complex propagation kernel [1]

bandlimit_border=(0.8, 1): applies a smooth bandlimit cut-off instead of hard-edge.

Examples

julia> field = zeros(ComplexF32, (4,4)); field[3,3] = 1
1
julia> as, t = angular_spectrum(field, 100e-9, 632e-9, 10e-6)
(ComplexF32[6.519258f-8 - 3.5390258f-7im -2.5097302f-7 + 1.1966712f-6im 0.00041754358f0 - 0.00027736276f0im -2.5097302f-7 + 1.1966712f-6im; -2.5136978f-7 + 1.1966767f-6im 8.540863f-7 - 4.0512546f-6im -0.001423778f0 + 0.00093840604f0im 8.540863f-7 - 4.0512546f-6im; 0.00041754544f0 - 0.0002773702f0im -0.0014237723f0 + 0.0009384034f0im 0.5497784f0 + 0.8353027f0im -0.0014237723f0 + 0.0009384034f0im; -2.5136978f-7 + 1.1966767f-6im 8.540863f-7 - 4.0512546f-6im -0.001423778f0 + 0.00093840604f0im 8.540863f-7 - 4.0512546f-6im], (H = ComplexF32[0.54519475f0 + 0.8383094f0im 0.5456109f0 + 0.8380386f0im … 0.54685986f0 + 0.8372242f0im 0.5456109f0 + 0.8380386f0im; 0.5456109f0 + 0.8380386f0im 0.5460271f0 + 0.83776754f0im … 0.54727626f0 + 0.83695203f0im 0.5460271f0 + 0.83776754f0im; … ; 0.54685986f0 + 0.8372242f0im 0.54727626f0 + 0.83695203f0im … 0.548526f0 + 0.8361335f0im 0.54727626f0 + 0.83695203f0im; 0.5456109f0 + 0.8380386f0im 0.5460271f0 + 0.83776754f0im … 0.54727626f0 + 0.83695203f0im 0.5460271f0 + 0.83776754f0im], L = 1.0e-5))

References

Matsushima, Kyoji, and Tomoyoshi Shimobaba. "Band-limited angular spectrum method for numerical simulation of free-space propagation in far and near fields." Optics express 17.22 (2009): 19662-19673.